requirements.txt
pytest==8.1.1
h11==0.14.0
pip==23.2.1
attrs==22.2.0
distro==1.9.0
wheel==0.41.2
Jinja2==3.1.4
setuptools==68.2.0
pytz==2024.2
typing_extensions==4.10.0
trio==0.24.0
outcome==1.3.0.post0
sniffio==1.3.1
sortedcontainers==2.4.0
cffi==1.16.0
idna==3.6
mysql-connector-python==9.2.0
numpy==2.1.1
execnet==2.1.1
Appium-Python-Client==1.3.0
selenium==3.141.0
urllib3==1.26.18
click==8.1.7
flake8==7.1.1
pyflakes==3.2.0
pycodestyle==2.12.1
pandas==2.2.3
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
pluggy==1.4.0
iniconfig==2.0.0
certifi==2024.2.2
six==1.16.0
pycparser==2.21
websocket-client==1.8.0
eyes-selenium==5.26.0
marshmallow==2.21.0
packaging==24.0
wsproto==1.2.0
requests==2.31.0
python-dotenv==1.0.1